Put things in their right place
